location hist + some functions

+++ src/Config.h
@@ -7,15 +7,15 @@
 
 class Config {
 public:
-  int room_id;
-  int temp_min;
-  int temp_max;
-  int hum_max;
-  int lum_min;
-  int infra_min;
-  std::string time_change_conf;
+    float room_id;
+    float temp_min;
+    float temp_max;
+    float hum_max;
+    float lum_min;
+    float infra_min;
+    std::string time_change_conf;
 
-  bool updateConfigs(User user, int room_id, std::string timestamp, int);
+    bool updateConfigs(User user, int room_id, std::string timestamp, int);
 };
